Making use of tenantId in most sql queries so that one tenant can not modify another tenant

+++ modules/components/org.wso2.appcloud.core/src/main/java/org/wso2/appcloud/core/SQLQueryConstants.java
@@ -106,46 +106,46 @@ public class SQLQueryConstants {
             " ON app.id = icon.application_id WHERE app.tenant_id=?";
 
     public static final String GET_VERSION_LIST_OF_APPLICATION =
-            "SELECT name FROM AC_VERSION WHERE application_id = (SELECT id FROM AC_APPLICATION WHERE hash_id=?)";
+            "SELECT name FROM AC_VERSION WHERE application_id = (SELECT id FROM AC_APPLICATION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_VERSION_HASH_IDS_OF_APPLICATION =
-            "SELECT hash_id FROM AC_VERSION WHERE application_id = (SELECT id FROM AC_APPLICATION WHERE hash_id=?)";
+            "SELECT hash_id FROM AC_VERSION WHERE application_id = (SELECT id FROM AC_APPLICATION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_VERSION_HASH_IDS_OF_APPLICATION_BY_VERSION_HASH_ID =
-            "SELECT hash_id FROM AC_VERSION WHERE application_id = (SELECT application_id FROM AC_VERSION WHERE hash_id=?)";
+            "SELECT hash_id FROM AC_VERSION WHERE application_id = (SELECT application_id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_APPLICATION_HASH_ID_BY_VERSION_HASH_ID =
-            "SELECT hash_id FROM AC_APPLICATION WHERE id = (SELECT application_id FROM AC_VERSION WHERE hash_id=?)";
+            "SELECT hash_id FROM AC_APPLICATION WHERE id = (SELECT application_id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_APPLICATION_BY_HASH_ID =
             "SELECT app.*, type.name as app_type_name, icon.icon as icon FROM AC_APPLICATION app JOIN AC_APP_TYPE type " +
-            "ON app.app_type_id = type.id JOIN AC_APP_ICON icon ON app.id = icon.application_id WHERE app.hash_id=?";
+            "ON app.app_type_id = type.id JOIN AC_APP_ICON icon ON app.id = icon.application_id WHERE app.hash_id=? AND app.tenant_id=?";
 
     public static final String GET_ALL_VERSIONS_OF_APPLICATION =
             "SELECT version.*, runtime.name as runtime_name, runtime.id as runtime_id FROM AC_VERSION version JOIN " +
             "AC_RUNTIME runtime ON version.runtime_id = runtime.id WHERE version.application_id = (SELECT id FROM " +
-            "AC_APPLICATION WHERE hash_id=?)";
+            "AC_APPLICATION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_APPLICATION_NAME_BY_HASH_ID =
-            "SELECT name FROM AC_APPLICATION WHERE hash_id = ?";
+            "SELECT name FROM AC_APPLICATION WHERE hash_id = ? AND tenant_id=?";
 
     public static final String GET_APPLICATION_HASH_ID_BY_NAME =
             "SELECT hash_id FROM AC_APPLICATION WHERE name=? AND tenant_id=?";
 
     public static final String GET_APPLICATION_ID =
-            "SELECT id FROM AC_APPLICATION WHERE hash_id=?";
+            "SELECT id FROM AC_APPLICATION WHERE hash_id=? AND tenant_id=?";
 
     public static final String GET_VERSION_ID =
-            "SELECT id FROM AC_VERSION WHERE hash_id=?";
+            "SELECT id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?";
 
     public static final String GET_ALL_TAGS_OF_VERSION =
-            "SELECT * FROM AC_TAG WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?)";
+            "SELECT * FROM AC_TAG WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_ALL_RUNTIME_PROPERTIES_OF_VERSION =
-            "SELECT * FROM AC_RUNTIME_PROPERTY WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?)";
+            "SELECT * FROM AC_RUNTIME_PROPERTY WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_CONTAINER_SERVICE_PROXIES =
-            "SELECT * FROM AC_CONTAINER_SERVICE_PROXY WHERE container_id=?";
+            "SELECT * FROM AC_CONTAINER_SERVICE_PROXY WHERE container_id=? AND tenant_id=?";
 
     public static final String GET_TRANSPORTS_FOR_RUNTIME =
             "SELECT name, port, protocol, service_prefix FROM AC_TRANSPORT WHERE id IN (SELECT transport_id FROM AC_RUNTIME_TRANSPORT " +
@@ -161,14 +161,14 @@ public class SQLQueryConstants {
             "SELECT * FROM AC_RUNTIME WHERE id = ?";
 
     public static final String GET_ALL_EVENTS_OF_APPLICATION =
-            "select * from AC_EVENT A where A.version_id = (SELECT id FROM AC_VERSION WHERE hash_id=?) and A.id >= " +
+            "select * from AC_EVENT A where A.version_id = (SELECT id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?) and A.id >= " +
             "(select MAX(B.id) from AC_EVENT B where B.version_id = A.version_id and B.name = A.name)";
 
     public static final String GET_DEPLOYMENT =
-            "SELECT * from AC_DEPLOYMENT where id=(SELECT deployment_id from AC_VERSION WHERE hash_id=?)";
+            "SELECT * from AC_DEPLOYMENT where id=(SELECT deployment_id from AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String GET_CONTAINER =
-            "SELECT * FROM AC_CONTAINER WHERE deployment_id=?";
+            "SELECT * FROM AC_CONTAINER WHERE deployment_id=? AND tenant_id=?";
 
     public static final String GET_CONTAINER_SERVICE_PROXY = "SELECT AC_CONTAINER_SERVICE_PROXY.name, " +
             "AC_CONTAINER_SERVICE_PROXY.protocol, AC_CONTAINER_SERVICE_PROXY.port, " +
@@ -176,7 +176,7 @@ public class SQLQueryConstants {
             "FROM AC_CONTAINER_SERVICE_PROXY " +
             "INNER JOIN AC_CONTAINER ON AC_CONTAINER_SERVICE_PROXY.container_id = AC_CONTAINER.id " +
             "INNER JOIN AC_DEPLOYMENT ON AC_CONTAINER.deployment_id = AC_DEPLOYMENT.id " +
-            "INNER JOIN AC_VERSION ON AC_DEPLOYMENT.id = AC_VERSION.deployment_id WHERE AC_VERSION.hash_id=?";
+            "INNER JOIN AC_VERSION ON AC_DEPLOYMENT.id = AC_VERSION.deployment_id WHERE AC_VERSION.hash_id=? AND tenant_id=?";
 
 
     public static final String GET_ALL_APP_VERSIONS_CREATED_BEFORE_X_DAYS_AND_NOT_WHITE_LISTED =
@@ -195,59 +195,59 @@ public class SQLQueryConstants {
 
     public static final String UPDATE_RUNTIME_PROPERTIES =
             "UPDATE AC_RUNTIME_PROPERTY SET name=?, value=? WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?)" +
-            " AND name=?";
+            " AND name=? AND tenant_id=?";
 
     public static final String UPDATE_VERSION_WITH_DEPLOYMENT =
             "UPDATE AC_VERSION SET deployment_id=? WHERE hash_id=? AND tenant_id=?";
 
     public static final String UPDATE_TAG =
-            "UPDATE AC_TAG SET name=?, value=? WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?) AND name=? ";
+            "UPDATE AC_TAG SET name=?, value=? WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?) AND name=? AND tenant_id=?";
 
     public static final String UPDATE_APPLICATION_STATUS =
-            "UPDATE AC_VERSION SET status=? WHERE hash_id=?";
+            "UPDATE AC_VERSION SET status=? WHERE hash_id=? AND tenant_id=?";
 
     public static final String UPDATE_CONTAINER_SERVICE_PROXY = "UPDATE AC_CONTAINER_SERVICE_PROXY " +
             "INNER JOIN AC_CONTAINER ON AC_CONTAINER_SERVICE_PROXY.container_id = AC_CONTAINER.id " +
             "INNER JOIN AC_DEPLOYMENT ON AC_CONTAINER.deployment_id = AC_DEPLOYMENT.id " +
             "INNER JOIN AC_VERSION ON AC_DEPLOYMENT.id = AC_VERSION.deployment_id " +
             "SET AC_CONTAINER_SERVICE_PROXY.host_url=? " +
-            "WHERE AC_VERSION.hash_id=?";
+            "WHERE AC_VERSION.hash_id=? AND AC_VERSION.tenant_id=?";
 
     public static final String  UPDATE_APPLICATION_DEFAULT_VERSION = "UPDATE AC_APPLICATION " +
-            "SET default_version=? WHERE hash_id=?";
+            "SET default_version=? WHERE hash_id=? AND tenant_id=?";
 
 	public static final String  UPDATE_WHITE_LIST_APPLICATION_VERSION = "UPDATE AC_VERSION " +
-	                                                                    "SET is_white_listed=? WHERE hash_id=?";
+	                                                                    "SET is_white_listed=? WHERE hash_id=? AND tenant_id=?";
 
     public static final String UPDATE_APP_VERSION_CON_SPEC = "UPDATE AC_VERSION SET con_spec_cpu = ?, " +
-            "con_spec_memory = ? WHERE hash_id = ?";
+            "con_spec_memory = ? WHERE hash_id = ? AND tenant_id=?";
 
 
 
     /*Delete Queries*/
 
     public static final String DELETE_RUNTIME_PROPERTY =
-            "DELETE FROM AC_RUNTIME_PROPERTY WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?) AND name=?";
+            "DELETE FROM AC_RUNTIME_PROPERTY WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?) AND name=? AND tenant_id=?";
 
     public static final String DELETE_TAG =
-            "DELETE FROM AC_TAG WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?) AND name=?";
+            "DELETE FROM AC_TAG WHERE version_id=(SELECT id FROM AC_VERSION WHERE hash_id=?) AND name=? AND tenant_id=?";
 
-    public static final String DELETE_APPLICATION = "DELETE FROM AC_APPLICATION WHERE hash_id=?";
+    public static final String DELETE_APPLICATION = "DELETE FROM AC_APPLICATION WHERE hash_id=? AND tenant_id=?";
 
-    public static final String DELETE_VERSION = "DELETE FROM AC_VERSION WHERE hash_id=?";
+    public static final String DELETE_VERSION = "DELETE FROM AC_VERSION WHERE hash_id=? AND tenant_id=?";
 
     public static final String DELETE_VERSIONS_OF_APPLICATION =
-            "DELETE FROM AC_VERSION WHERE application_id = (SELECT id FROM AC_APPLICATION WHERE hash_id=?)";
+            "DELETE FROM AC_VERSION WHERE application_id = (SELECT id FROM AC_APPLICATION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String DELETE_DEPLOYMENT =
-            "DELETE FROM AC_DEPLOYMENT WHERE id=(SELECT deployment_id FROM AC_VERSION WHERE hash_id=?)";
+            "DELETE FROM AC_DEPLOYMENT WHERE id=(SELECT deployment_id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
     public static final String DELETE_ALL_DEPLOYMENT_OF_APPLICATION =
             "DELETE FROM AC_DEPLOYMENT WHERE id in (SELECT deployment_id from AC_VERSION WHERE application_id = " +
-            "(SELECT id FROM AC_APPLICATION WHERE hash_id=?))";
+            "(SELECT id FROM AC_APPLICATION WHERE hash_id=? AND tenant_id=?))";
 
     public static final String DELETE_ALL_APP_VERSION_EVENTS =
-            "Delete from AC_EVENT where version_id = (SELECT id FROM AC_VERSION WHERE hash_id=?)";
+            "Delete from AC_EVENT where version_id = (SELECT id FROM AC_VERSION WHERE hash_id=? AND tenant_id=?)";
 
 	public static final String GET_TENANT_APPLICATION_COUNT = "SELECT COUNT(*) FROM AC_APPLICATION WHERE tenant_id = ?";
 }
